--- examples/bring_your_own_fips202/mlkem_native/debug
@@ -1,1 +1,0 @@
-../../../mlkem/debug
