+++ examples/bring_your_own_fips202/mlkem_native/debug.c
@@ -1,0 +1,1 @@
+../../../mlkem/debug.c
