+++ examples/bring_your_own_fips202/mlkem_native/debug.h
@@ -1,0 +1,1 @@
+../../../mlkem/debug.h
