--- examples/custom_backend/mlkem_native/mlkem/debug
@@ -1,1 +1,0 @@
-../../../../mlkem/debug
