+++ examples/custom_backend/mlkem_native/mlkem/debug.c
@@ -1,0 +1,1 @@
+../../../../mlkem/debug.c
