+++ mlkem/indcpa.c
@@ -17,7 +17,7 @@
 #include "symmetric.h"
 
 #include "arith_backend.h"
-#include "debug/debug.h"
+#include "debug.h"
 
 #include "cbmc.h"
