+++ mlkem/reduce.h
@@ -8,7 +8,7 @@
 #include <stdint.h>
 #include "cbmc.h"
 #include "common.h"
-#include "debug/debug.h"
+#include "debug.h"
 
 /* Static namespacing
  * This is to facilitate building multiple instances
